+++ docs/mkdocs/docs/api/arctic_uri.md
@@ -23,7 +23,7 @@ Available options for S3:
 | secret                | S3 secret access key                                                                                                                                            |
 | path_prefix           | Path within S3 bucket to use for data storage                                                                                                                   |
 | aws_auth              | AWS authentication method. If setting is `default` (or `true` for backward compatibility), authentication to endpoint will be computed via [AWS default credential provider chain](https://docs.aws.amazon.com/sdk-for-cpp/v1/developer-guide/credproviders.html). If setting is `sts`, AWS Security Token Service (STS) will be the authentication method used. If no options are provided AWS authentication will not be used and you should specify access and secret in the URI. More info about `sts` is provided [here](https://docs.arcticdb.io/latest/aws/#aws-security-token-service-sts-setup) |
-| aws_profile           | Named AWS profile (from the shared config/credentials files) to authenticate with. Requires `aws_auth` to be set. With `aws_auth=default` the profile is resolved by the default credentials provider chain; with `aws_auth=sts` it is used by AWS Security Token Service (STS). More info about `sts` is provided [here](https://docs.arcticdb.io/latest/aws/#aws-security-token-service-sts-setup) |
+| aws_profile           | Named AWS profile (from the shared config/credentials files) to authenticate with. Requires `aws_auth` to be set. With `aws_auth=default` the profile is resolved by the default credentials provider chain. If using conda release with `aws-sdk-cpp<=1.11.747`, only profile set in AWS credetial files will be loaded; With `aws_auth=sts` it is used by AWS Security Token Service (STS). More info about `sts` is provided [here](https://docs.arcticdb.io/latest/aws/#aws-security-token-service-sts-setup) |
 
 Note: When connecting to AWS, `region` can be automatically deduced from the endpoint if the given endpoint
 specifies the region and `region` is not set.
